change blit prototype

+++ trunk/ibitmapm.h
@@ -92,7 +92,7 @@ int ibitmap_clip(const int *clipdst, const int *clipsrc, int *x, int *y,
  * sx, sy    - source rectangle position in source bitmap
  * w, h      - source rectangle width and height in source bitmap
  * mask      - coloekey of source bitmap, ignored if no IBLIT_MASK
- * mode      - blit flag bits (IBLIT_CLIP, IBLIT_MASK)
+ * mode      - flags of IBLIT_CLIP, IBLIT_MASK, IBLIT_HFLIP, IBLIT_VFLIP...
  */
 int ibitmap_blit(struct IBITMAP *dest, int dx, int dy, struct IBITMAP *src, 
           int x, int y, int w, int h, unsigned long mask, int mode);
@@ -113,13 +113,13 @@ int ibitmap_fill(struct IBITMAP *dst, int dx, int dy, int w, int h,
 /**********************************************************************
  * Bitmap Basic Interface
  **********************************************************************/
-typedef int (*ibitmap_blitter_norm)(char *, long, const char *, 
-              int, int, long, int, long);
-typedef int (*ibitmap_blitter_mask)(char *, long, const char *, 
-              int, int, long, int, long, unsigned long);
+typedef int (*ibitmap_blitter_norm)(char *, long, const char *, long, 
+              int, int, int, long);
+typedef int (*ibitmap_blitter_mask)(char *, long, const char *, long,
+              int, int, int, long, unsigned long);
 
-typedef int (*ibitmap_blitter_flip)(char *, long, const char *,
-              int, int, long, int, long, unsigned long, int);
+typedef int (*ibitmap_blitter_flip)(char *, long, const char *, long,
+              int, int, int, long, unsigned long, int);
 
 typedef int (*ibitmap_filler)(char *, long, int, int, int, unsigned long);
 
@@ -139,10 +139,10 @@ typedef int (*ibitmap_filler)(char *, long, int, int, int, unsigned long);
  * ibitmap_funcset - set basic bitmap functions, returns zero for successful
  * mode - function index, there are two blitters, (0/1 normal/transparent) 
  * proc - extra function pointer, set it to null to use default method:
- *   int imyblitter0(char *dst, long pitch1, const char *src, 
- *       int w, int h, long pitch2, int pixelbyte);
- *   int imyblitter1(char *dst, long pitch1, const char *src, 
- *       int w, int h, long pitch2, int pixelbyte, unsigned long ckey);
+ *   int imyblitter0(char *dst, long pitch1, const char *src, long pitch2,
+ *       int w, int h, int pixelbyte, long linesize);
+ *   int imyblitter1(char *dst, long pitch1, const char *src, long pitch2
+ *       int w, int h, int pixelbyte, unsigned long mask, long linesize);
  * there are two blitters, normal blitter and transparent blitter, which 
  * ibitmap_blit() will call to complete blit operation. 
  */
